pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dwellio"
version = "0.1.0"
description = "Dwellio backend and ETL foundation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.3,<3.0",
  "email-validator>=2.2,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "PyYAML>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.4,<25.0",
  "mypy>=1.10,<2.0",
  "pytest>=8.2,<9.0",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.5,<1.0",
  "httpx>=0.27,<1.0",
]

[project.scripts]
dwellio-migrate = "infra.scripts.run_migrations:main"
dwellio-job = "app.jobs.cli:main"

[tool.setuptools.packages.find]
include = ["app*", "infra*"]
exclude = ["apps*", "docs*", "sql*", "supabase*", "tests*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_configs = true
warn_return_any = true
check_untyped_defs = true
no_implicit_optional = true
